+++ lizard_datasource/scripts.py
@@ -46,6 +46,10 @@ def cache_latest_values(ds):
             properties.DATA_CAN_HAVE_VALUE_LAYER_SCRIPT)):
         return  # For now, we don't know what to do in this case
 
+    # Only actually do something if the script is due.
+    if not ds.activation_for_cache_script():
+        return
+
     for layer in _yield_layers(ds):
         # This creates the datasource layer in the database, if it
         # didn't exist yet